requirements.txt
google-genai==1.28.0
dotenv==0.9.9